requirements.txt
PyYAML>=6.0
requests>=2.28.0
urllib3>=1.26.0
tqdm>=4.64.0
